hode: fix Task.run not being correctly restored after mstCollide
This is causing issues in level 6 'pwr2' screen 3, where level was
not restarted when Andy got splashed against the wall.

The monster task was set to sm_wait with a delay = 1000. Because,
the task.run would reset to execute the next instruction, it would
loop back to re-execute the specialAction opcode and change
Andy animation.

1448 (CC): special_action code:10 num:100
144C (00): sm_wait delay:1000
1450 (E5): jmp_imm offset:0x1480